+++ src/blocks/xrandr.rs
@@ -49,7 +49,7 @@ impl Monitor {
 }
 
 pub struct Xrandr {
-    id: u64,
+    id: usize,
     text: ButtonWidget,
     update_interval: Duration,
     monitors: Vec<Monitor>,
@@ -236,7 +236,7 @@ impl ConfigBlock for Xrandr {
     type Config = XrandrConfig;
 
     fn new(
-        id: u64,
+        id: usize,
         block_config: Self::Config,
         config: Config,
         _tx_update_request: Sender<Task>,
@@ -312,7 +312,7 @@ impl Block for Xrandr {
         Ok(())
     }
 
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
